config: make it clear how and where we write global configuration
fixes: https://github.com/planetscale/project-big-bang/issues/158
